+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This is used for recording Invader's changes. This changelog is based on
 ### Added
 - invader-edit-qt: Added font previewing
 
+### Fixed
+- invader-compare: Fixed crashing on maps with stubbed or invalid tags
+
 ## [0.33.0] - 2020-04-05
 ### Added
 - New tool: invader-compare: Compares tags between maps and/or tags directories