+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 This is used for recording Invader's changes. This changelog is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.33.1] - 2020-04-06
 ### Added
 - invader-edit-qt: Added font previewing